Extract HandState and HandStateExtensions to separate package
